using go-tasks to run log retrieval on schedule

+++ scripts/docker-entrypoint.sh
@@ -9,6 +9,7 @@ set -o pipefail
 : ${CF_AUTH_KEY:?required}
 : ${CF_LOGS_DIRECTORY:='/logs'}
 
+: ${FILEBEAT_USER:=filebeat} # set during docker build
 : ${FILEBEAT_CONFIG:='/opt/filebeat/filebeat.yml'}
 : ${INDEX_TEMPLATE_FILE:='/opt/filebeat/index-template.json.tpl'}
 : ${SAMPLE_RATE:='0.01'}
@@ -23,15 +24,6 @@ export TZ='UTC'
 export CF_AUTH_EMAIL CF_AUTH_KEY CF_ZONE_ID CF_LOGS_DIRECTORY
 export FILEBEAT_CONFIG ES_INDEX_JSON_ENABLED
 export ES_INDEX ES_INDEX_SHARDS ES_INDEX_REPLICAS ES_INDEX_REFRESH
-
-if [[ ! -r "${FILEBEAT_CONFIG}" ]]; then
-  echo >&2 "ERROR: unable to read filebeat config at \"${FILEBEAT_CONFIG}\", exiting.."
-  exit 1
-fi
-
-if [[ ! -d "${CF_LOGS_DIRECTORY}" ]]; then
-  mkdir "${CF_LOGS_DIRECTORY}"
-fi
 
 generate_index_template() {
   jq \
@@ -122,10 +114,16 @@ cat <<EOM
 EOM
 }
 
+setup_cron() {
+cat >/etc/crontabs/${FILEBEAT_USER} <<EOF
+*/5 * * * *  /usr/local/bin/docker-entrypoint.sh --from-cron
+EOF
+}
+
 get_logs() {
   local _end_utc="$(date '+%R %D' -d '1 minute ago')"
   local _end_epoch="$(date +%s -d "${_end_utc}")"
-  local _start_epoch="$((${_end_epoch}-380))"
+  local _start_epoch="$((${_end_epoch}-360))"
   local _suffix="$(date +%F_%H:%M:%S -d "@${_start_epoch}")"
   /usr/local/bin/get_cloudflare_logs.sh \
     --access-logs \
@@ -136,45 +134,58 @@ get_logs() {
     > ${CF_LOGS_DIRECTORY}/cloudflare_${_suffix}.log
 }
 
-if [[ "${ES_INDEX_JSON_ENABLED}" == "true" ]]; then
-  echo >&2 '## generating index template'
+if [[ "${1:-}" == "--from-cron" ]]; then
+
+  set +e ## tolerate errors from this point on
+  get_logs
+  find ${CF_LOGS_DIRECTORY} -type f -name "cloudflare_*.log" -mmin +5 -delete
+
+else
+
+  if [[ ! -r "${FILEBEAT_CONFIG}" ]]; then
+    echo >&2 "ERROR: unable to read filebeat config at \"${FILEBEAT_CONFIG}\", exiting.."
+    exit 1
+  fi
+
+  if [[ ! -d "${CF_LOGS_DIRECTORY}" ]]; then
+    mkdir "${CF_LOGS_DIRECTORY}"
+  fi
+
+  echo >&2 '## generating crontab entry'
   echo >&2
-  generate_index_template
+  setup_cron
+
+  if [[ "${ES_INDEX_JSON_ENABLED}" == "true" ]]; then
+    echo >&2 '## generating index template'
+    echo >&2
+    generate_index_template
+  fi
+
+  echo >&2 '## running Filebeat setup'
+  echo >&2
+  filebeat \
+    -c "${FILEBEAT_CONFIG}" \
+    -E setup.ilm.enabled=true \
+    -E setup.ilm.overwrite=true \
+    -E setup.template.enabled=true \
+    -E setup.template.overwrite=true \
+    -E setup.template.json.enabled=${ES_INDEX_JSON_ENABLED} \
+    setup --index-management
+
+  echo >&2
+  echo >&2 '## initializing index'
+  init_message | jq -cM > ${CF_LOGS_DIRECTORY}/cloudflare_init.log
+
+  echo >&2
+  echo >&2 '## launching Filebeat'
+  echo >&2
+  filebeat -c "${FILEBEAT_CONFIG}" &
+
+  sleep 10
+
+  echo >&2
+  echo >&2 '## launching process retrieving logs from Cloudflare..'
+  echo >&2
+  go-tasks --allow-unprivileged
 fi
 
-echo >&2 '## running Filebeat setup'
-echo >&2
-filebeat \
-  -c "${FILEBEAT_CONFIG}" \
-  -E setup.ilm.enabled=true \
-  -E setup.ilm.overwrite=true \
-  -E setup.template.enabled=true \
-  -E setup.template.overwrite=true \
-  -E setup.template.json.enabled=${ES_INDEX_JSON_ENABLED} \
-  setup --index-management
-
-echo >&2
-echo >&2 '## initializing index'
-init_message | jq -cM > ${CF_LOGS_DIRECTORY}/cloudflare_init.log
-
-echo >&2
-echo >&2 '## launching Filebeat'
-echo >&2
-filebeat -c "${FILEBEAT_CONFIG}" &
-
-sleep 10
-
-set +e ## tolerate errors from this point on
-
-echo >&2
-echo >&2 '## retrieving logs from Cloudflare..'
-echo >&2
-while true
-do
-  get_logs &
-  find ${CF_LOGS_DIRECTORY} -type f -name "cloudflare_*.log" -mmin +5 -delete
-  echo >&2
-  echo >&2 '## Waiting 5 minutes, before getting another chunk of logs..'
-  echo >&2
-  sleep 300
-done
